Refactor transaction tests into several modules
The `test_transactions.py` was getting huge.
